DAOS-6063 control: tidy prior to event notify feature
Split out improvements from code review of PR 3999:
- separate pci and time related helpers into their own files
- document vfio disable (server config file) in admin guide
