+++ src/control/cmd/dmg/pretty/storage_nvme.go
@@ -32,6 +32,7 @@ import (
 
 	"github.com/dustin/go-humanize"
 
+	"github.com/daos-stack/daos/src/control/common"
 	"github.com/daos-stack/daos/src/control/lib/control"
 	"github.com/daos-stack/daos/src/control/lib/txtfmt"
 	"github.com/daos-stack/daos/src/control/server/storage"
@@ -48,6 +49,13 @@ func printNvmeControllerSummary(nvme *storage.NvmeController, out io.Writer, opt
 	return w.Err
 }
 
+func getTimestampString(secs uint64) string {
+	if secs == 0 {
+		return "N/A"
+	}
+	return common.FormatTime(time.Unix(int64(secs), 0))
+}
+
 func printNvmeHealth(stat *storage.NvmeHealth, out io.Writer, opts ...PrintConfigOption) error {
 	w := txtfmt.NewErrWriter(out)
 
@@ -61,8 +69,7 @@ func printNvmeHealth(stat *storage.NvmeHealth, out io.Writer, opts ...PrintConfi
 	iw := txtfmt.NewIndentWriter(out)
 
 	if stat.Timestamp > 0 {
-		fmt.Fprintf(iw, "Timestamp:%s\n",
-			time.Unix(int64(stat.Timestamp), 0).Format(time.UnixDate))
+		fmt.Fprintf(iw, "Timestamp:%s\n", getTimestampString(stat.Timestamp))
 	}
 
 	fmt.Fprintf(iw, "Temperature:%dK(%.02fC)\n", stat.TempK(), stat.TempC())
