+++ src/control/cmd/dmg/pretty/storage_nvme_test.go
@@ -49,7 +49,7 @@ func TestPretty_PrintNVMeHealthMap(t *testing.T) {
 		t.Fatal(err)
 	}
 	controllerAwTS.HealthStats.Timestamp = tt
-	ttStr := time.Unix(int64(tt), 0).Format(time.UnixDate)
+	ttStr := getTimestampString(tt)
 
 	for name, tc := range map[string]struct {
 		hsm         control.HostStorageMap
