+++ src/control/common/env_utils.go
@@ -1,5 +1,5 @@
 //
-// (C) Copyright 2019 Intel Corporation.
+// (C) Copyright 2019-2020 Intel Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -20,6 +20,7 @@
 // Any reproduction of computer software, computer software documentation, or
 // portions thereof marked with this legend must also reproduce the markings.
 //
+
 package common
 
 import (
@@ -54,6 +55,8 @@ func ScrubEnvironmentExcept(whitelist []string) {
 	}
 }
 
+// DisableProxyScrubEnv disabled removal of proxy variables from the process
+// environment.
 const DisableProxyScrubEnv = "DAOS_DISABLE_PROXY_SCRUB"
 
 func proxyScrubIsDisabled() bool {
