+++ src/control/common/pci_utils.go
@@ -1,0 +1,56 @@
+//
+// (C) Copyright 2020 Intel Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// GOVERNMENT LICENSE RIGHTS-OPEN SOURCE SOFTWARE
+// The Government's rights to use, modify, reproduce, release, perform, display,
+// or disclose this software are subject to the terms of the Apache License as
+// provided in Contract No. 8F-30005.
+// Any reproduction of computer software, computer software documentation, or
+// portions thereof marked with this legend must also reproduce the markings.
+//
+
+package common
+
+import (
+	"strconv"
+	"strings"
+
+	"github.com/pkg/errors"
+)
+
+// ParsePCIAddress returns separated components of BDF format PCI address.
+func ParsePCIAddress(addr string) (dom, bus, dev, fun uint64, err error) {
+	parts := strings.Split(addr, ":")
+	devFunc := strings.Split(parts[len(parts)-1], ".")
+	if len(parts) != 3 || len(devFunc) != 2 {
+		err = errors.Errorf("unexpected pci address bdf format: %q", addr)
+		return
+	}
+
+	if dom, err = strconv.ParseUint(parts[0], 16, 64); err != nil {
+		return
+	}
+	if bus, err = strconv.ParseUint(parts[1], 16, 32); err != nil {
+		return
+	}
+	if dev, err = strconv.ParseUint(devFunc[0], 16, 32); err != nil {
+		return
+	}
+	if fun, err = strconv.ParseUint(devFunc[1], 16, 32); err != nil {
+		return
+	}
+
+	return
+}
