+++ src/control/common/proto/mocks.go
@@ -24,17 +24,9 @@
 package proto
 
 import (
-	"context"
-	"fmt"
-
-	"google.golang.org/grpc"
-
-	"github.com/daos-stack/daos/src/control/common"
 	ctlpb "github.com/daos-stack/daos/src/control/common/proto/ctl"
 	mgmtpb "github.com/daos-stack/daos/src/control/common/proto/mgmt"
 	"github.com/daos-stack/daos/src/control/server/storage"
-	"github.com/daos-stack/daos/src/control/system"
-	"github.com/pkg/errors"
 )
 
 // MockNvmeNamespace is a mock protobuf Namespace message used in tests for
@@ -133,182 +125,4 @@ func MockScmMountPoint(varIdx ...int32) *ctlpb.ScmNamespace_Mount {
 var MockPoolList = []*mgmtpb.ListPoolsResp_Pool{
 	{Uuid: "12345678-1234-1234-1234-123456789abc", Svcreps: []uint32{1, 2}},
 	{Uuid: "12345678-1234-1234-1234-cba987654321", Svcreps: []uint32{0}},
-}
-
-type MockMgmtSvcClientConfig struct {
-	ListPoolsRet      *common.MockListPoolsResult
-	KillErr           error
-	PoolQueryResult   *mgmtpb.PoolQueryResp
-	PoolQueryErr      error
-	PoolSetPropResult *mgmtpb.PoolSetPropResp
-	PoolSetPropErr    error
-}
-
-type MockMgmtSvcClient struct {
-	Cfg   MockMgmtSvcClientConfig
-	Calls []string
-}
-
-func NewMockMgmtSvcClient(cfg MockMgmtSvcClientConfig) mgmtpb.MgmtSvcClient {
-	return &MockMgmtSvcClient{Cfg: cfg}
-}
-
-func (m *MockMgmtSvcClient) PoolCreate(ctx context.Context, req *mgmtpb.PoolCreateReq, o ...grpc.CallOption) (*mgmtpb.PoolCreateResp, error) {
-	// return successful pool creation results
-	// initialize with zero values indicating mgmt.CTL_SUCCESS
-	return &mgmtpb.PoolCreateResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) PoolResolveID(ctx context.Context, req *mgmtpb.PoolResolveIDReq, o ...grpc.CallOption) (*mgmtpb.PoolResolveIDResp, error) {
-	return &mgmtpb.PoolResolveIDResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) PoolDestroy(ctx context.Context, req *mgmtpb.PoolDestroyReq, o ...grpc.CallOption) (*mgmtpb.PoolDestroyResp, error) {
-	// return successful pool destroy results
-	// initialize with zero values indicating mgmt.CTL_SUCCESS
-	return &mgmtpb.PoolDestroyResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) PoolExtend(ctx context.Context, req *mgmtpb.PoolExtendReq, o ...grpc.CallOption) (*mgmtpb.PoolExtendResp, error) {
-	return &mgmtpb.PoolExtendResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) PoolEvict(ctx context.Context, req *mgmtpb.PoolEvictReq, o ...grpc.CallOption) (*mgmtpb.PoolEvictResp, error) {
-	return &mgmtpb.PoolEvictResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) PoolReintegrate(ctx context.Context, req *mgmtpb.PoolReintegrateReq, o ...grpc.CallOption) (*mgmtpb.PoolReintegrateResp, error) {
-	// return successful pool reintegrate results
-	// initialize with zero values indicating mgmt.CTL_SUCCESS
-	return &mgmtpb.PoolReintegrateResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) PoolExclude(ctx context.Context, req *mgmtpb.PoolExcludeReq, o ...grpc.CallOption) (*mgmtpb.PoolExcludeResp, error) {
-	// return successful pool Exclude results
-	// initialize with zero values indicating mgmt.CTL_SUCCESS
-	return &mgmtpb.PoolExcludeResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) PoolDrain(ctx context.Context, req *mgmtpb.PoolDrainReq, o ...grpc.CallOption) (*mgmtpb.PoolDrainResp, error) {
-	return &mgmtpb.PoolDrainResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) PoolQuery(ctx context.Context, req *mgmtpb.PoolQueryReq, _ ...grpc.CallOption) (*mgmtpb.PoolQueryResp, error) {
-	if m.Cfg.PoolQueryErr != nil {
-		return nil, m.Cfg.PoolQueryErr
-	}
-	return m.Cfg.PoolQueryResult, nil
-}
-
-func (m *MockMgmtSvcClient) PoolSetProp(ctx context.Context, req *mgmtpb.PoolSetPropReq, _ ...grpc.CallOption) (*mgmtpb.PoolSetPropResp, error) {
-	if m.Cfg.PoolSetPropErr != nil {
-		return nil, m.Cfg.PoolSetPropErr
-	}
-	return m.Cfg.PoolSetPropResult, nil
-}
-
-// returnACLResult returns the mock ACL results - either an error or an ACLResp
-func (m *MockMgmtSvcClient) returnACLResult() (*mgmtpb.ACLResp, error) {
-	return nil, nil
-}
-
-func (m *MockMgmtSvcClient) PoolGetACL(ctx context.Context, req *mgmtpb.GetACLReq, o ...grpc.CallOption) (*mgmtpb.ACLResp, error) {
-	return m.returnACLResult()
-}
-
-func (m *MockMgmtSvcClient) PoolOverwriteACL(ctx context.Context, req *mgmtpb.ModifyACLReq, o ...grpc.CallOption) (*mgmtpb.ACLResp, error) {
-	return m.returnACLResult()
-}
-
-func (m *MockMgmtSvcClient) PoolUpdateACL(ctx context.Context, req *mgmtpb.ModifyACLReq, o ...grpc.CallOption) (*mgmtpb.ACLResp, error) {
-	return m.returnACLResult()
-}
-
-func (m *MockMgmtSvcClient) PoolDeleteACL(ctx context.Context, req *mgmtpb.DeleteACLReq, o ...grpc.CallOption) (*mgmtpb.ACLResp, error) {
-	return m.returnACLResult()
-}
-
-func (m *MockMgmtSvcClient) BioHealthQuery(ctx context.Context, req *mgmtpb.BioHealthReq, o ...grpc.CallOption) (*mgmtpb.BioHealthResp, error) {
-
-	// return successful bio health results
-	// initialize with zero values indicating mgmt.CTL_SUCCESS
-	return &mgmtpb.BioHealthResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) SmdQuery(ctx context.Context, req *mgmtpb.SmdQueryReq, o ...grpc.CallOption) (*mgmtpb.SmdQueryResp, error) {
-	// return successful device state
-	// initialize with zero values indicating mgmt.CTRL_SUCCESS
-	return &mgmtpb.SmdQueryResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) StorageSetFaulty(ctx context.Context, req *mgmtpb.DevStateReq, o ...grpc.CallOption) (*mgmtpb.DevStateResp, error) {
-
-	// return suscessful FAULTY device state
-	// initialize with zero values indicating mgmt.CTRL_SUCCESS
-	return &mgmtpb.DevStateResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) Join(ctx context.Context, req *mgmtpb.JoinReq, o ...grpc.CallOption) (*mgmtpb.JoinResp, error) {
-	m.Calls = append(m.Calls, fmt.Sprintf("Join %d", req.GetRank()))
-
-	// if req rank == NilRank, map uuid to rank
-	// otherwise, simulate join where requested rank is valid and returned
-	rank := req.GetRank()
-	nilRank := system.NilRank
-	if nilRank.Uint32() == rank {
-		switch req.GetUuid() {
-		case common.MockUUID(0):
-			rank = 0
-		case common.MockUUID(1):
-			rank = 1
-		default:
-			return nil, errors.New("Mock Join doesn't recognise UUID")
-		}
-	}
-
-	return &mgmtpb.JoinResp{Rank: rank}, nil
-}
-
-func (m *MockMgmtSvcClient) GetAttachInfo(ctx context.Context, in *mgmtpb.GetAttachInfoReq, opts ...grpc.CallOption) (*mgmtpb.GetAttachInfoResp, error) {
-	return &mgmtpb.GetAttachInfoResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) PrepShutdownRanks(ctx context.Context, req *mgmtpb.RanksReq, o ...grpc.CallOption) (*mgmtpb.RanksResp, error) {
-	return &mgmtpb.RanksResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) StopRanks(ctx context.Context, req *mgmtpb.RanksReq, o ...grpc.CallOption) (*mgmtpb.RanksResp, error) {
-	return &mgmtpb.RanksResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) PingRanks(ctx context.Context, req *mgmtpb.RanksReq, o ...grpc.CallOption) (*mgmtpb.RanksResp, error) {
-	return &mgmtpb.RanksResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) ResetFormatRanks(ctx context.Context, req *mgmtpb.RanksReq, o ...grpc.CallOption) (*mgmtpb.RanksResp, error) {
-	return &mgmtpb.RanksResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) StartRanks(ctx context.Context, req *mgmtpb.RanksReq, o ...grpc.CallOption) (*mgmtpb.RanksResp, error) {
-	return &mgmtpb.RanksResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) ListPools(ctx context.Context, req *mgmtpb.ListPoolsReq, o ...grpc.CallOption) (*mgmtpb.ListPoolsResp, error) {
-	if m.Cfg.ListPoolsRet.Err != nil {
-		return nil, m.Cfg.ListPoolsRet.Err
-	}
-	return &mgmtpb.ListPoolsResp{Pools: MockPoolList, Status: m.Cfg.ListPoolsRet.Status}, nil
-}
-
-func (m *MockMgmtSvcClient) LeaderQuery(ctx context.Context, req *mgmtpb.LeaderQueryReq, _ ...grpc.CallOption) (*mgmtpb.LeaderQueryResp, error) {
-	return &mgmtpb.LeaderQueryResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) ListContainers(ctx context.Context, req *mgmtpb.ListContReq, o ...grpc.CallOption) (*mgmtpb.ListContResp, error) {
-	// return successful list containers results
-	return &mgmtpb.ListContResp{}, nil
-}
-
-func (m *MockMgmtSvcClient) ContSetOwner(ctx context.Context, req *mgmtpb.ContSetOwnerReq, o ...grpc.CallOption) (*mgmtpb.ContSetOwnerResp, error) {
-	return nil, nil
 }
