+++ src/control/server/instance_exec.go
@@ -31,9 +31,7 @@ import (
 
 	"github.com/daos-stack/daos/src/control/build"
 	srvpb "github.com/daos-stack/daos/src/control/common/proto/srv"
-	"github.com/daos-stack/daos/src/control/server/config"
 	"github.com/daos-stack/daos/src/control/server/ioserver"
-	"github.com/daos-stack/daos/src/control/system"
 )
 
 // IOServerRunner defines an interface for starting and stopping the
@@ -145,7 +143,7 @@ func (srv *IOServerInstance) exit(exitErr error) {
 // will only return (if no errors are returned during setup) on IO server
 // process exit (triggered by harness shutdown through context cancellation
 // or abnormal IO server process termination).
-func (srv *IOServerInstance) run(ctx context.Context, membership *system.Membership, recreateSBs bool) (err error) {
+func (srv *IOServerInstance) run(ctx context.Context, recreateSBs bool) (err error) {
 	errChan := make(chan error)
 
 	if err = srv.format(ctx, recreateSBs); err != nil {
@@ -166,7 +164,7 @@ func (srv *IOServerInstance) run(ctx context.Context, membership *system.Members
 
 // Run is the processing loop for an IOServerInstance. Starts are triggered by
 // receiving true on instance start channel.
-func (srv *IOServerInstance) Run(ctx context.Context, membership *system.Membership, cfg *config.Server) {
+func (srv *IOServerInstance) Run(ctx context.Context, recreateSBs bool) {
 	for {
 		select {
 		case <-ctx.Done():
@@ -175,7 +173,7 @@ func (srv *IOServerInstance) Run(ctx context.Context, membership *system.Members
 			if !relaunch {
 				return
 			}
-			srv.exit(srv.run(ctx, membership, cfg.RecreateSuperblocks))
+			srv.exit(srv.run(ctx, recreateSBs))
 		}
 	}
 }
